Moved the watcher into the fswatch lib
