+++ nagios/nagios.go
@@ -5,14 +5,11 @@ import (
 	"fmt"
 	"log"
 	"os"
-	"path"
 	"strconv"
 	"strings"
 	"time"
 	"unicode"
 
-	watch "github.com/bensallen/sqlios/fswatch"
-	"github.com/fsnotify/fsnotify"
 	"github.com/influxdata/influxdb/client/v2"
 	"github.com/jprichardson/readline-go"
 )
@@ -342,40 +339,4 @@ func Reader(blockc chan Block, filec chan *os.File, endOfFile chan bool, errc ch
 		}
 	}
 
-}
-
-// Watcher watches the input file using inotify, knotifyd, etc for CREATE events.
-// Nagios does atomic updates of status.dat by writting out to a temporary file,
-// removing status.dat and moving the temporary file to status.dat. The last
-// event seen in this process is a CREATE, so we use that to kick off reading.
-// Watcher waits on the done chan forever.
-func Watcher(input *string, filec chan *os.File, done chan bool, errc chan error) {
-
-	path := path.Dir(*input)
-
-	dir, err := os.Stat(path)
-
-	if err != nil || !dir.IsDir() {
-		log.Fatalf("Parent directory of input: %s, could not stat or is not a directory, watcher bailing!", path)
-	}
-
-	var eventc = make(chan *fsnotify.Event, 128)
-
-	go func() {
-		for event := range eventc {
-			// TODO: What does this notation actually mean?
-			if event.Op&fsnotify.Create == fsnotify.Create {
-				file, err := os.Open(*input)
-				if err != nil {
-					errc <- err
-				}
-				filec <- file
-			}
-		}
-	}()
-
-	watch.Watch(path, *input, eventc, errc)
-
-	//Wait until someone tells us we're done
-	<-done
 }
